// src/tv-api/types/episodes.ts
import type { EpisodeLink, Image, Rating } from "./shared";

export interface Episode {
  id: number;
  url: string;
  name: string;
  season: number;
  number: number;
  type: string;
  airdate: string;
  airtime: string;
  airstamp: string;
  runtime: number;
  rating: Rating;
  image: Image;
  summary: string;
  _links: EpisodeLink;
}